+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+In next release...
+
+Backwards incompatibilities
+###########################
+
+- The ``route`` method was renamed to ``connect``.
+
 1.0 (2009-11-12)
 ----------------